+++ authservice/Dockerfile
@@ -17,4 +17,4 @@ COPY . .
 EXPOSE 8002
 
 # Define the command to run your app
-CMD ["node", "service.js"]
+CMD ["node", "auth-service.js"]
