+++ authservice/auth-service.test.js
@@ -20,7 +20,7 @@ beforeAll(async () => {
   mongoServer = await MongoMemoryServer.create();
   const mongoUri = mongoServer.getUri();
   process.env.MONGODB_URI = mongoUri;
-  app = require('./service'); 
+  app = require('./auth-service'); 
 });
 
 afterAll(async () => {
