+++ gatewayservice/gateway-service.js
