+++ gatewayservice/gateway-service.test.js
@@ -1,6 +1,6 @@
 const request = require('supertest');
 const axios = require('axios');
-const app = require('./service'); 
+const app = require('./gateway-service'); 
 
 afterAll(async () => {
     app.close();
