+++ userservice/user-service.test.js
@@ -8,7 +8,7 @@ beforeAll(async () => {
   mongoServer = await MongoMemoryServer.create();
   const mongoUri = mongoServer.getUri();
   process.env.MONGODB_URI = mongoUri;
-  app = require('./service'); 
+  app = require('./user-service'); 
 });
 
 afterAll(async () => {
